+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/RoomDetailFragment.kt
@@ -65,6 +65,7 @@ import im.vector.matrix.android.api.session.room.send.SendState
 import im.vector.matrix.android.api.session.room.timeline.TimelineEvent
 import im.vector.matrix.android.api.session.room.timeline.getLastMessageContent
 import im.vector.matrix.android.api.session.room.timeline.getTextEditableContent
+import im.vector.matrix.android.api.session.sync.SyncState
 import im.vector.matrix.android.api.session.user.model.User
 import im.vector.riotx.R
 import im.vector.riotx.core.di.ScreenComponent
@@ -185,6 +186,7 @@ class RoomDetailFragment :
     private lateinit var scrollOnNewMessageCallback: ScrollOnNewMessageCallback
     private lateinit var scrollOnHighlightedEventCallback: ScrollOnHighlightedEventCallback
     @Inject lateinit var eventHtmlRenderer: EventHtmlRenderer
+    @Inject lateinit var vectorPreferences: VectorPreferences
 
 
     override fun getLayoutResId() = R.layout.fragment_room_detail
@@ -246,6 +248,14 @@ class RoomDetailFragment :
                 is SendMode.QUOTE -> enterSpecialMode(mode.timelineEvent, R.drawable.ic_quote, false)
                 is SendMode.REPLY -> enterSpecialMode(mode.timelineEvent, R.drawable.ic_reply, false)
             }
+        }
+
+        roomDetailViewModel.selectSubscribe(RoomDetailViewState::syncState) { syncState ->
+            syncProgressBar.visibility = when (syncState) {
+                is SyncState.RUNNING -> if (syncState.afterPause) View.VISIBLE else View.GONE
+                else                 -> View.GONE
+            }
+            syncProgressBarWrap.visibility = syncProgressBar.visibility
         }
     }
 
@@ -390,7 +400,7 @@ class RoomDetailFragment :
         recyclerView.setController(timelineEventController)
         timelineEventController.callback = this
 
-        if (VectorPreferences.swipeToReplyIsEnabled(requireContext())) {
+        if (vectorPreferences.swipeToReplyIsEnabled()) {
             val swipeCallback = RoomMessageTouchHelperCallback(requireContext(),
                                                                R.drawable.ic_reply,
                                                                object : RoomMessageTouchHelperCallback.QuickReplayHandler {
@@ -483,7 +493,7 @@ class RoomDetailFragment :
         composerLayout.sendButton.setOnClickListener {
             val textMessage = composerLayout.composerEditText.text.toString()
             if (textMessage.isNotBlank()) {
-                roomDetailViewModel.process(RoomDetailActions.SendMessage(textMessage, VectorPreferences.isMarkdownEnabled(requireContext())))
+                roomDetailViewModel.process(RoomDetailActions.SendMessage(textMessage, vectorPreferences.isMarkdownEnabled()))
             }
         }
         composerLayout.composerRelatedMessageCloseButton.setOnClickListener {
@@ -508,7 +518,7 @@ class RoomDetailFragment :
             items.add(DialogListItem.SendFile)
             // Send voice
 
-            if (VectorPreferences.isSendVoiceFeatureEnabled(this)) {
+            if (vectorPreferences.isSendVoiceFeatureEnabled()) {
                 items.add(DialogListItem.SendVoice.INSTANCE)
             }
 
@@ -517,7 +527,7 @@ class RoomDetailFragment :
             //items.add(DialogListItem.SendSticker)
             // Camera
 
-            //if (VectorPreferences.useNativeCamera(this)) {
+            //if (vectorPreferences.useNativeCamera()) {
             items.add(DialogListItem.TakePhoto)
             items.add(DialogListItem.TakeVideo)
             //} else {
@@ -639,8 +649,12 @@ class RoomDetailFragment :
 
     private fun renderSendMessageResult(sendMessageResult: SendMessageResult) {
         when (sendMessageResult) {
-            is SendMessageResult.MessageSent,
+            is SendMessageResult.MessageSent                -> {
+                // Clear composer
+                composerLayout.composerEditText.text = null
+            }
             is SendMessageResult.SlashCommandHandled        -> {
+                sendMessageResult.messageRes?.let { showSnackWithMessage(getString(it)) }
                 // Clear composer
                 composerLayout.composerEditText.text = null
             }
@@ -965,7 +979,7 @@ class RoomDetailFragment :
 //                vibrate = true
             }
 
-//            if (vibrate && VectorPreferences.vibrateWhenMentioning(context)) {
+//            if (vibrate && vectorPreferences.vibrateWhenMentioning()) {
 //                val v= context.getSystemService(Context.VIBRATOR_SERVICE) as? Vibrator
 //                if (v?.hasVibrator() == true) {
 //                    v.vibrate(100)